refactor: split test_kg_families_e2e into three focused holistic modules
The 661-line module was three suites stapled together. The CASES
registry, vocabularies, and discovery helpers now live in the shared
_family_cases.py; test_kg_usage_smoke.py drives the all-family sweep,
test_kg_registry_integrity.py holds the CASES-vs-discovery gates, and
test_kg_catalog_declarations.py holds the source-slot gate plus the
printed asymmetry catalog. References in docs and docstrings updated.
