+++ AGENTS.md
@@ -67,7 +67,7 @@ When helping with FeatureGroups, ComputeFrameworks, or Extenders, leverage these
 - **Type hints**: use modern forms (`list[str]`, `dict[str, int]`, `X | None`).
 - **Formatting**: ruff format with line length 120.
 - **Tests**: every new feature or bug fix must come with tests; follow the patterns in the existing `open_kgo/.../tests/` trees.
-- **Test contract harness (in-repo, not a published library)**: KG connectors are tested via a shared, inherited contract suite, the same pattern mloda-registry uses, but it currently lives **inside this repo** and is **not** packaged or published for external reuse (unlike mloda-registry's separate `mloda-testing` package). Structure: a universal `KgConnectorContractBase` in `open_kgo/feature_groups/kg/tests/kg_contract.py`, a per-family `<Family>ContractTestBase` in each `open_kgo/feature_groups/kg/<family>/tests/kg_<family>_contract.py`, the shared `run_query`/`make_valid_credentials` helpers in `open_kgo/feature_groups/kg/tests/_helpers.py` (reader discovery in `_discovery.py`), and the resource-cache loaders in the production module `open_kgo/feature_groups/kg/fixtures.py` (caches cleared per-test by `open_kgo/feature_groups/kg/conftest.py`). A holistic all-family smoke plus cross-family gap diagnostics live in `open_kgo/feature_groups/kg/tests/test_kg_families_e2e.py`. Each concrete plugin's test class subclasses its family base and supplies a few small adapter methods. When adding a connector, extend the existing family base rather than writing standalone tests. Note also that, unlike mloda-registry, there is no framework-adapter-mixin layer and no single canonical fixture dataset.
+- **Test contract harness (in-repo, not a published library)**: KG connectors are tested via a shared, inherited contract suite, the same pattern mloda-registry uses, but it currently lives **inside this repo** and is **not** packaged or published for external reuse (unlike mloda-registry's separate `mloda-testing` package). Structure: a universal `KgConnectorContractBase` in `open_kgo/feature_groups/kg/tests/kg_contract.py`, a per-family `<Family>ContractTestBase` in each `open_kgo/feature_groups/kg/<family>/tests/kg_<family>_contract.py`, the shared `run_query`/`make_valid_credentials` helpers in `open_kgo/feature_groups/kg/tests/_helpers.py` (reader discovery in `_discovery.py`), and the resource-cache loaders in the production module `open_kgo/feature_groups/kg/fixtures.py` (caches cleared per-test by `open_kgo/feature_groups/kg/conftest.py`). A holistic all-family smoke plus cross-family gap diagnostics live in `open_kgo/feature_groups/kg/tests/test_kg_usage_smoke.py`, `test_kg_registry_integrity.py`, and `test_kg_catalog_declarations.py`, all over the shared `_family_cases.py` registry. Each concrete plugin's test class subclasses its family base and supplies a few small adapter methods. When adding a connector, extend the existing family base rather than writing standalone tests. Note also that, unlike mloda-registry, there is no framework-adapter-mixin layer and no single canonical fixture dataset.
 - **Supply chain**: `[tool.uv] exclude-newer = "7 days"` in `pyproject.toml` defers new dependency releases by 7 days; `exclude-newer-package` pins mloda, mloda-testing, and mloda-registry to a longer window. Do not edit these without a reason.
 - **Commits**: use [Conventional Commits](https://www.conventionalcommits.org/) (`feat:`, `fix:`, `chore:`, `docs:`, `test:`, `refactor:`, `style:`, `ci:`, `build:`, `perf:`). semantic-release computes the next version: `feat:` triggers a minor bump, all other types trigger a patch bump (see `.releaserc.yaml`).
